requirements.txt
beautifulsoup4
discord.py
psycopg[binary]
Requests